pyproject.toml
[tool.poetry]
name = "shopify-api_inventory_manager"
version = "0.2.0"
description = "A project description here."
authors = ["Craig Van <mail@craigvan.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
shopify-api = "^0.7.1"  # Example version, adjust as necessary
pandas = "^1.2.3"
prophet = "^1.0"
numpy = "^1.19.2"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"